backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
python-jose[cryptography]==3.3.0
